+++ code/webapp/src/components/attendance/EmployeeAttendanceCard.tsx
@@ -13,23 +13,7 @@ import type {
     AttendancePhase,
     TodayAttendanceEmployee,
 } from '@/types/attendance'
-
-// ── Phase Helpers ──────────────────────────────────────────────────────────────
-
-export function getPhaseCardClass(phase: AttendancePhase): string {
-    switch (phase) {
-        case 'pending':
-            return 'border-muted/60 opacity-70'
-        case 'checked-in':
-            return 'border-blue-200 dark:border-blue-800'
-        case 'at-lunch':
-            return 'border-orange-200 dark:border-orange-800'
-        case 'returned':
-            return 'border-teal-200 dark:border-teal-800'
-        case 'done':
-            return 'border-green-200 dark:border-green-800'
-    }
-}
+import { getPhaseCardClass } from './attendance-helpers'
 
 // ── Sub-components ─────────────────────────────────────────────────────────────
